+++ scripts/_helpers.py
@@ -4,14 +4,8 @@
 
 # coding: utf-8
 
-import os
 import re
 from pathlib import Path
-
-import numpy as np
-import pandas as pd
-import snakemake as sm
-from snakemake.script import Snakemake
 
 
 class Dict(dict):
@@ -94,6 +88,19 @@ def mock_snakemake(
         keyword arguments fixing the wildcards. Only necessary if wildcards are
         needed.
     """
+    import os
+
+    import snakemake as sm
+    from snakemake.api import Workflow
+    from snakemake.common import SNAKEFILE_CHOICES
+    from snakemake.script import Snakemake
+    from snakemake.settings.types import (
+        ConfigSettings,
+        DAGSettings,
+        ResourceSettings,
+        StorageSettings,
+        WorkflowSettings,
+    )
 
     script_dir = Path(__file__).parent.resolve()
     if root_dir is None:
@@ -113,7 +120,7 @@ def mock_snakemake(
             f" {root_dir} or scripts directory {script_dir}"
         )
     try:
-        for p in sm.SNAKEFILE_CHOICES:
+        for p in SNAKEFILE_CHOICES:
             if os.path.exists(p):
                 snakefile = p
                 break
@@ -122,12 +129,12 @@ def mock_snakemake(
         elif isinstance(configfiles, str):
             configfiles = [configfiles]
 
-        resource_settings = sm.ResourceSettings()
-        config_settings = sm.ConfigSettings(configfiles=map(Path, configfiles))
-        workflow_settings = sm.WorkflowSettings()
-        storage_settings = sm.StorageSettings()
-        dag_settings = sm.DAGSettings(rerun_triggers=[])
-        workflow = sm.Workflow(
+        resource_settings = ResourceSettings()
+        config_settings = ConfigSettings(configfiles=map(Path, configfiles))
+        workflow_settings = WorkflowSettings()
+        storage_settings = StorageSettings()
+        dag_settings = DAGSettings(rerun_triggers=[])
+        workflow = Workflow(
             config_settings,
             resource_settings,
             workflow_settings,